Break everything out of App except the base initialization for XAML. AppLogic is the new home.
