+++ src/cascadia/WindowsTerminal/AppHost.h
@@ -23,11 +23,12 @@ private:
 
     std::unique_ptr<IslandWindow> _window;
     winrt::TerminalApp::App _app;
+    winrt::TerminalApp::AppLogic _logic;
 
     void _HandleCreateWindow(const HWND hwnd, RECT proposedRect, winrt::TerminalApp::LaunchMode& launchMode);
     void _UpdateTitleBarContent(const winrt::Windows::Foundation::IInspectable& sender,
                                 const winrt::Windows::UI::Xaml::UIElement& arg);
-    void _UpdateTheme(const winrt::TerminalApp::App&,
+    void _UpdateTheme(const winrt::Windows::Foundation::IInspectable&,
                       const winrt::Windows::UI::Xaml::ElementTheme& arg);
     void _ToggleFullscreen(const winrt::Windows::Foundation::IInspectable& sender,
                            const winrt::TerminalApp::ToggleFullscreenEventArgs& arg);
